+++ snapshot-tests-core/src/test/java/de/skuzzle/test/snapshots/SnapshotTestOptionsTest.java
@@ -12,4 +12,10 @@ public class SnapshotTestOptionsTest {
         final SnapshotTestOptions annotation = SnapshotTestOptionsTest.class.getAnnotation(SnapshotTestOptions.class);
         assertThat(annotation.normalizeLineEndings()).isEqualTo(SnapshotTestOptions.DEFAULT_NORMALIZE_LINE_ENDINGS);
     }
+
+    @Test
+    void testCorrectDefaultValueForDiffFormat() {
+        final SnapshotTestOptions annotation = SnapshotTestOptionsTest.class.getAnnotation(SnapshotTestOptions.class);
+        assertThat(annotation.diffFormat()).isEqualTo(SnapshotTestOptions.DEFAULT_DIFF_FORMAT);
+    }
 }